+++ src/headless/types/plugins/mam/utils.d.ts
@@ -1,7 +1,8 @@
 /**
+ * @param {Element|Error} e
  * @param {Element} iq
  */
-export function onMAMError(iq: Element): Promise<void>;
+export function onMAMError(e: Element | Error, iq: Element): Promise<void>;
 /**
  * Handle returned IQ stanza containing Message Archive
  * Management (XEP-0313) preferences.
@@ -24,31 +25,36 @@ export function getMAMPrefsFromFeature(feature: Model): void;
 /**
  * @param {MUC} muc
  */
-export function preMUCJoinMAMFetch(muc: MUC): void;
+export function preMUCJoinMAMFetch(muc: MUC): Promise<void>;
 /**
  * @param {ChatBox|MUC} model
  * @param {Object} result
  * @param {Object} query
  * @param {Object} options
- * @param {('forwards'|'backwards'|null)} [should_page=null]
+ * @param {('forwards'|'backwards'|false)} [should_page=false]
  */
-export function handleMAMResult(model: ChatBox | MUC, result: any, query: any, options: any, should_page?: ("forwards" | "backwards" | null)): Promise<void>;
+export function handleMAMResult(model: ChatBox | MUC, result: any, query: any, options: any, should_page?: ("forwards" | "backwards" | false)): Promise<void>;
 /**
  * Fetch XEP-0313 archived messages based on the passed in criteria.
  * @param {ChatBox|MUC} model
- * @param {import('./types').MAMOptions} [options]
- * @param {('forwards'|'backwards'|null)} [should_page=null] - Determines whether
+ * @param {import('./types').FetchArchivedMessagesOptions} [options]
+ * @param {('forwards'|'backwards'|false)} [should_page=false] - Determines whether
  *  this function should recursively page through the entire result set if a limited
  *  number of results were returned.
  */
-export function fetchArchivedMessages(model: ChatBox | MUC, options?: import("./types").MAMOptions, should_page?: ("forwards" | "backwards" | null)): Promise<void>;
+export function fetchArchivedMessages(model: ChatBox | MUC, options?: import("./types").FetchArchivedMessagesOptions, should_page?: ("forwards" | "backwards" | false)): Promise<void>;
+/**
+ * Creates a placeholder to fetch messages at the top of the chat history.
+ * @param {ChatBox|MUC} model
+ */
+export function createScrollupPlaceholder(model: ChatBox | MUC): void;
 /**
  * Fetches messages that might have been archived *after*
  * the last archived message in our local cache.
  * @param {ChatBox|MUC} model
  */
-export function fetchNewestMessages(model: ChatBox | MUC): void;
-export type MUC = import("../muc/muc.js").default;
-export type ChatBox = import("../chat/model.js").default;
+export function fetchNewestMessages(model: ChatBox | MUC): Promise<void>;
+export type MUC = import("../muc/muc").default;
+export type ChatBox = import("../chat/model").default;
 export type Model = import("@converse/skeletor/src/types/helpers.js").Model;
 //# sourceMappingURL=utils.d.ts.map